+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This file is similar to the format suggested by [Keep a CHANGELOG](https://githu
 - [Patch] Add hover and focus differentiations for `unstyled` **Buttons** ([1326](https://github.com/optimizely/oui/pull/1326))
 - [Patch] Adjust spacing between drag handle and order number for **Tokens** ([#1328](https://github.com/optimizely/oui/pull/1328))
 - [Patch] Fix the **SearchPicker** results count ([#1330](https://github.com/optimizely/oui/pull/1330))
+- [Patch] Throw an error in SearchPicker's searchFunction Promise chain if the response is not an array. This will prevent a non-Array value from being set as defaultResults ([1325](https://github.com/optimizely/oui/pull/1325))
 
 ## 46.1.0 - 2020-05-08
 - [Feature] Add ability to specify a left or right icon for **Input** ([#1320](https://github.com/optimizely/oui/pull/1320))